Start refactoring config
