+++ utils/prompt.js
@@ -1,0 +1,9 @@
+var prompt = require('prompt');
+
+prompt.message = 'Input';
+
+prompt.start({
+  noHandleSIGINT: true,
+});
+
+module.exports = prompt;
